Use the same collapse method for top-down and bottom-up methods.
Previously top-down methods gave up after the first bad split was
proposed. This should be better.

Still have to get rid of the weird evaluation metric where both children
of a node are evaluated as splits, and both have to be valid. It should
just be: per-node, are the children of this node supported by a split.